Some changes according to the W3C structure for better security
In the past version I was just decoding the "access-token" in the AuthServiceProvider.php but Hash is the more secure than encoding/decoding. So I have used "Argon2 algorithm" which is most strong hash algorithm of current time.  And Encoded token will be decoded in the controller for making any changes in the server's application data. So there are 2 layer of security of different types. And after each login the private key will be different.So encoded token is hard to decode .