Some Changes according to W3C Protocol standards
in my previous version I was using a singla route with POST method for a full CRUD operation. Than I found this is not a good practice . So I did some research and understand the proper cause . So now I changed my code and divide the actions according to the propber HTTP request methods.
https://www.w3.org/Protocols/rfc2616/rfc2616-sec9.html